+++ shared/src/main/scala/com/cibo/evilplot/plot/renderers/TickRenderer.scala
@@ -32,6 +32,7 @@ package com.cibo.evilplot.plot.renderers
 
 import com.cibo.evilplot.geometry._
 import com.cibo.evilplot.plot.aesthetics.Theme
+import com.cibo.evilplot.plot.components.Position
 
 trait TickRenderer {
   def render(label: String): Drawable
@@ -48,49 +49,86 @@ object TickRenderer {
     * @param thickness  The thickness of the tick line.
     * @param rotateText The rotation of the label.
     */
+  @deprecated("Use axisTickRenderer", "EvilPlot 0.3.4")
   def xAxisTickRenderer(
     length: Double = defaultTickLength,
     thickness: Double = defaultTickThickness,
     rotateText: Double = 0
-  )(implicit theme: Theme): TickRenderer = new TickRenderer {
-    private def align(tick: Drawable, label: Drawable): Drawable =
-      if (rotateText.toInt % 90 == 0) Align.center(tick, label).reduce(_ above _)
-      else tick.above(label).transX(label.extent.width)
-
-    def render(label: String): Drawable = {
-      val line = Line(length, thickness).rotated(90)
-      align(
-        line.colored(theme.colors.tickLabel),
-        Style(
-          Text(label.toString, size = theme.fonts.tickLabelSize, fontFace = theme.fonts.fontFace),
-          theme.colors.tickLabel)
-          .rotated(rotateText)
-          .padTop(2)
-      )
-    }
-  }
+  )(implicit theme: Theme): TickRenderer = axisTickRenderer(
+    Position.Bottom,
+    length,
+    thickness,
+    rotateText
+  )
 
   /** Create a renderer to render a tick on the y axis.
     *
     * @param length    The length of the tick line.
     * @param thickness The thickness of the tick line.
     */
+  @deprecated("Use axisTickRenderer", "EvilPlot 0.3.4")
   def yAxisTickRenderer(
     length: Double = defaultTickLength,
     thickness: Double = defaultTickThickness
+  )(implicit theme: Theme): TickRenderer = axisTickRenderer(
+    Position.Left,
+    length,
+    thickness,
+    0
+  )
+
+  //TODO revisit alignment for rotated text labels so that tick is aligned to "highest" part of text
+  //     e.g. left-side for 1-89 & 181 - 269, right-side for 91 - 179 & 271-359
+  //     or expose alignment?
+  /** Create a renderer to render a tick on an arbitrary axis.
+    *
+    * @param position   The side of the plot where this axis will be added.
+    * @param length     The length of the tick line.
+    * @param thickness  The thickness of the tick line.
+    * @param rotateText The rotation of the label.
+    */
+  def axisTickRenderer(
+    position: Position,
+    length: Double = defaultTickLength,
+    thickness: Double = defaultTickThickness,
+    rotateText: Double = 0
   )(implicit theme: Theme): TickRenderer = new TickRenderer {
     def render(label: String): Drawable = {
-      val line = Line(length, thickness)
-      Align
-        .middle(
-          Style(
-            Text(label.toString, size = theme.fonts.tickLabelSize, fontFace = theme.fonts.fontFace),
-            theme.colors.tickLabel)
-            .padRight(2)
-            .padBottom(2),
-          line.colored(theme.colors.tickLabel)
-        )
-        .reduce(beside)
+      val line = Line(length, thickness).colored(theme.colors.tickLabel)
+      val verticalLine = line.rotated(90)
+      val text = Style(
+        Text(label.toString, size = theme.fonts.tickLabelSize, fontFace = theme.fonts.fontFace),
+        theme.colors.tickLabel
+      ).rotated(rotateText)
+      if (rotateText.toInt % 90 == 0) {
+        position match {
+          case Position.Left =>
+            Align.middle(text.padRight(2).padBottom(2), line).reduce(beside)
+          case Position.Right =>
+            Align.middle(line, text.padLeft(2).padBottom(2)).reduce(beside)
+          case Position.Bottom =>
+            Align.center(verticalLine, text.padTop(2)).reduce(_ above _)
+          case Position.Top =>
+            Align.center(text.padBottom(2), verticalLine).reduce(_ above _)
+          case _ =>
+            text
+        }
+      } else {
+        position match {
+          case Position.Left =>
+            val labelDrawable = text.padRight(2).padBottom(2)
+            labelDrawable.transY(-labelDrawable.extent.height).beside(line)
+          case Position.Right =>
+            val labelDrawable = text.padLeft(2).padBottom(2)
+            line.beside(labelDrawable.transY(-labelDrawable.extent.height))
+          case Position.Bottom =>
+            verticalLine.above(text.padTop(2)).transX(text.extent.width)
+          case Position.Top =>
+            verticalLine.below(text.padBottom(2)).transX(text.extent.width)
+          case _ =>
+            text
+        }
+      }
     }
   }
 }